Separate navigation tests
