+++ Tests/PlayerTests/BackwardNavigationTests.swift
@@ -11,7 +11,7 @@ import CoreMedia
 import Nimble
 import XCTest
 
-final class NavigationTests: XCTestCase {
+final class BackwardNavigationTests: XCTestCase {
     func testCanReturnForOnDemandAtBeginningWithoutPreviousItem() {
         let item = PlayerItem.simple(url: Stream.onDemand.url)
         let player = Player(item: item)
@@ -34,7 +34,7 @@ final class NavigationTests: XCTestCase {
         expect(player.streamType).toEventually(equal(.onDemand))
 
         waitUntil { done in
-            player.seek(to: CMTime(value: 5, timescale: 1), toleranceBefore: .zero, toleranceAfter: .zero) { finished in
+            player.seek(to: CMTime(value: 5, timescale: 1), toleranceBefore: .zero, toleranceAfter: .zero) { _ in
                 done()
             }
         }
@@ -90,66 +90,6 @@ final class NavigationTests: XCTestCase {
         expect(player.canReturnToPrevious()).to(beFalse())
     }
 
-    func testCanAdvanceForOnDemandWithNextItem() {
-        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
-        let item2 = PlayerItem.simple(url: Stream.live.url)
-        let player = Player(items: [item1, item2])
-        expect(player.streamType).toEventually(equal(.onDemand))
-        expect(player.canAdvanceToNext()).to(beTrue())
-    }
-
-    func testCannotAdvanceForOnDemandWithoutNextItem() {
-        let item = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(item: item)
-        expect(player.streamType).toEventually(equal(.onDemand))
-        expect(player.canAdvanceToNext()).to(beFalse())
-    }
-
-    func testCanAdvanceForLiveWithNextItem() {
-        let item1 = PlayerItem.simple(url: Stream.live.url)
-        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(items: [item1, item2])
-        expect(player.streamType).toEventually(equal(.live))
-        expect(player.canAdvanceToNext()).to(beTrue())
-    }
-
-    func testCannotAdvanceForLiveWithoutNextItem() {
-        let item = PlayerItem.simple(url: Stream.live.url)
-        let player = Player(item: item)
-        expect(player.streamType).toEventually(equal(.live))
-        expect(player.canAdvanceToNext()).to(beFalse())
-    }
-
-    func testCanAdvanceForDvrWithNextItem() {
-        let item1 = PlayerItem.simple(url: Stream.dvr.url)
-        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(items: [item1, item2])
-        expect(player.streamType).toEventually(equal(.dvr))
-        expect(player.canAdvanceToNext()).to(beTrue())
-    }
-
-    func testCannotAdvanceForDvrWithoutNextItem() {
-        let item = PlayerItem.simple(url: Stream.dvr.url)
-        let player = Player(item: item)
-        expect(player.streamType).toEventually(equal(.dvr))
-        expect(player.canAdvanceToNext()).to(beFalse())
-    }
-
-    func testCanAdvanceForUnknownWithNextItem() {
-        let item1 = PlayerItem.simple(url: Stream.unavailable.url)
-        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(items: [item1, item2])
-        expect(player.streamType).to(equal(.unknown))
-        expect(player.canAdvanceToNext()).to(beTrue())
-    }
-
-    func testCannotAdvanceForUnknownWithoutNextItem() {
-        let item = PlayerItem.simple(url: Stream.unavailable.url)
-        let player = Player(item: item)
-        expect(player.streamType).to(equal(.unknown))
-        expect(player.canAdvanceToNext()).to(beFalse())
-    }
-
     func testReturnForOnDemandAtBeginningWithoutPreviousItem() {
         let item = PlayerItem.simple(url: Stream.onDemand.url)
         let player = Player(item: item)
@@ -174,7 +114,7 @@ final class NavigationTests: XCTestCase {
         expect(player.streamType).toEventually(equal(.onDemand))
 
         waitUntil { done in
-            player.seek(to: CMTime(value: 5, timescale: 1), toleranceBefore: .zero, toleranceAfter: .zero) { finished in
+            player.seek(to: CMTime(value: 5, timescale: 1), toleranceBefore: .zero, toleranceAfter: .zero) { _ in
                 done()
             }
         }
@@ -234,74 +174,6 @@ final class NavigationTests: XCTestCase {
         let player = Player(item: item)
         expect(player.streamType).toEventually(equal(.unknown))
         player.returnToPreviousItem()
-        expect(player.currentIndex).to(equal(0))
-    }
-
-    func testAdvanceForOnDemandWithNextItem() {
-        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
-        let item2 = PlayerItem.simple(url: Stream.live.url)
-        let player = Player(items: [item1, item2])
-        expect(player.streamType).toEventually(equal(.onDemand))
-        player.advanceToNext()
-        expect(player.currentIndex).to(equal(1))
-    }
-
-    func testAdvanceForOnDemandWithoutNextItem() {
-        let item = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(item: item)
-        expect(player.streamType).toEventually(equal(.onDemand))
-        player.advanceToNext()
-        expect(player.currentIndex).to(equal(0))
-    }
-
-    func testAdvanceForLiveWithNextItem() {
-        let item1 = PlayerItem.simple(url: Stream.live.url)
-        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(items: [item1, item2])
-        expect(player.streamType).toEventually(equal(.live))
-        player.advanceToNext()
-        expect(player.currentIndex).to(equal(1))
-    }
-
-    func testAdvanceForLiveWithoutNextItem() {
-        let item = PlayerItem.simple(url: Stream.live.url)
-        let player = Player(item: item)
-        expect(player.streamType).toEventually(equal(.live))
-        player.advanceToNext()
-        expect(player.currentIndex).to(equal(0))
-    }
-
-    func testAdvanceForDvrWithNextItem() {
-        let item1 = PlayerItem.simple(url: Stream.dvr.url)
-        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(items: [item1, item2])
-        expect(player.streamType).toEventually(equal(.dvr))
-        player.advanceToNext()
-        expect(player.currentIndex).to(equal(1))
-    }
-
-    func testAdvanceForDvrWithoutNextItem() {
-        let item = PlayerItem.simple(url: Stream.dvr.url)
-        let player = Player(item: item)
-        expect(player.streamType).toEventually(equal(.dvr))
-        player.advanceToNext()
-        expect(player.currentIndex).to(equal(0))
-    }
-
-    func testAdvanceForUnknownWithNextItem() {
-        let item1 = PlayerItem.simple(url: Stream.unavailable.url)
-        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(items: [item1, item2])
-        expect(player.streamType).to(equal(.unknown))
-        player.advanceToNext()
-        expect(player.currentIndex).to(equal(1))
-    }
-
-    func testAdvanceForUnknownWithoutNextItem() {
-        let item = PlayerItem.simple(url: Stream.unavailable.url)
-        let player = Player(item: item)
-        expect(player.streamType).to(equal(.unknown))
-        player.advanceToNext()
         expect(player.currentIndex).to(equal(0))
     }
 }
