+++ Tests/PlayerTests/ItemForwardNavigationTests.swift
@@ -9,27 +9,7 @@
 import Nimble
 import XCTest
 
-final class ItemNavigationTests: XCTestCase {
-    func testCanReturnToPreviousItem() {
-        let item1 = PlayerItem.simple(url: Stream.item(numbered: 1).url)
-        let item2 = PlayerItem.simple(url: Stream.item(numbered: 2).url)
-        let player = Player(items: [item1, item2])
-        player.advanceToNextItem()
-        expect(player.canReturnToPreviousItem()).to(beTrue())
-    }
-
-    func testCannotReturnToPreviousItemAtFront() {
-        let item1 = PlayerItem.simple(url: Stream.item(numbered: 1).url)
-        let item2 = PlayerItem.simple(url: Stream.item(numbered: 2).url)
-        let player = Player(items: [item1, item2])
-        expect(player.canReturnToPreviousItem()).to(beFalse())
-    }
-
-    func testCannotReturnToPreviousItemWhenEmpty() {
-        let player = Player()
-        expect(player.canReturnToPreviousItem()).to(beFalse())
-    }
-
+final class ItemForwardNavigationTests: XCTestCase {
     func testCanAdvanceToNextItem() {
         let item1 = PlayerItem.simple(url: Stream.item(numbered: 1).url)
         let item2 = PlayerItem.simple(url: Stream.item(numbered: 2).url)
@@ -56,15 +36,6 @@ final class ItemNavigationTests: XCTestCase {
         let item3 = PlayerItem.simple(url: Stream.item(numbered: 3).url)
         let player = Player(items: [item1, item2, item3])
         player.advanceToNextItem()
-        expect(player.canReturnToPreviousItem()).toAlways(beTrue(), until: .milliseconds(500))
-    }
-
-    func testCanReturnToPreviousItemOnFailedItem() {
-        let item1 = PlayerItem.simple(url: Stream.item(numbered: 1).url)
-        let item2 = PlayerItem.simple(url: Stream.unavailable.url)
-        let item3 = PlayerItem.simple(url: Stream.item(numbered: 3).url)
-        let player = Player(items: [item1, item2, item3])
-        player.advanceToNextItem()
         expect(player.canAdvanceToNextItem()).toAlways(beTrue(), until: .milliseconds(500))
     }
 
@@ -83,33 +54,6 @@ final class ItemNavigationTests: XCTestCase {
         player.advanceToNextItem()
         player.advanceToNextItem()
         expect(player.currentIndex).to(equal(1))
-    }
-
-    func testReturnToPreviousItem() {
-        let item1 = PlayerItem.simple(url: Stream.item(numbered: 1).url)
-        let item2 = PlayerItem.simple(url: Stream.item(numbered: 2).url)
-        let player = Player(items: [item1, item2])
-        player.advanceToNextItem()
-        player.returnToPreviousItem()
-        expect(player.currentIndex).to(equal(0))
-    }
-
-    func testReturnToPreviousItemAtFront() {
-        let item1 = PlayerItem.simple(url: Stream.item(numbered: 1).url)
-        let item2 = PlayerItem.simple(url: Stream.item(numbered: 2).url)
-        let player = Player(items: [item1, item2])
-        player.returnToPreviousItem()
-        expect(player.currentIndex).to(equal(0))
-    }
-
-    func testReturnToPreviousItemOnFailedItem() {
-        let item1 = PlayerItem.simple(url: Stream.item(numbered: 1).url)
-        let item2 = PlayerItem.simple(url: Stream.unavailable.url)
-        let item3 = PlayerItem.simple(url: Stream.item(numbered: 3).url)
-        let player = Player(items: [item1, item2, item3])
-        player.advanceToNextItem()
-        player.returnToPreviousItem()
-        expect(player.currentIndex).to(equal(0))
     }
 
     func testAdvanceToNextItemOnFailedItem() {
